requirements.txt
streamlit==1.21.0
faiss-cpu==1.7.3
chromadb==0.3.25
requests==2.31.0
sentence-transformers==2.2.2
PyMuPDF==1.22.0
python-docx==0.8.11
pyyaml==6.0
dotenv==1.0.0
